server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
fastf1==3.4.0
anthropic==0.52.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
pydantic>=2.0,<3
